requirements.txt
onnxruntime
onnx
onnxslim
picamera2
opencv-python-headless
pandas
paho-mqtt
pyyaml
ultralytics